requirements.txt
asgiref==3.2.3
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.3
django-bootstrap3==12.0.3
django-ckeditor==5.9.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-modeltranslation==0.14.4
django-phonenumber-field==4.0.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
packaging==20.3
phonenumbers==8.11.5
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
pyparsing==2.4.6
python-decouple==3.3
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
